+++ testing/test_integration.py
@@ -10,8 +10,8 @@ import setuptools_scm._integration.setuptools
 from .wd_wrapper import WorkDir
 from setuptools_scm import PRETEND_KEY
 from setuptools_scm import PRETEND_KEY_NAMED
+from setuptools_scm._integration.setuptools import _warn_on_old_setuptools
 from setuptools_scm._run_cmd import run
-from setuptools_scm.integration import _warn_on_old_setuptools
 
 
 @pytest.fixture